Conflicts:
	assets/Shaders/batchfragmentshader.frag
	include/UtH/Engine/DefaultScene.hpp
	src/Engine/DefaultScene.cpp
	src/Engine/UtHEngine.cpp

+++ include/UtH/Engine/SceneManager.hpp
@@ -18,7 +18,7 @@ namespace uth
 		void GoToScene(int SceneNumber);
 		void GoToScene(Scene* newScene);
 
-		bool Update(double dt);
+		bool Update(float dt);
 		bool Draw();
